requirements.txt
django
djangorestframework
requests
pycryptodome
Flask
Flask-WTF